pyproject.toml
[project]
name = "michael-travel-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.0.5",
    "langgraph>=1.0.7",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.5",
]
